moved heightfield file; minor changes for warnings;
